:recycle: refactor(search): Refactor search into smaller components
